requirements.txt
setuptools
wheel
pyspark==2.4.5
pyarrow==0.17.1
mlflow==1.8.0
pytest==5.3.1
python-dotenv==0.10.3
databricks-cli==0.10.0
pyyaml>=5.1
scikit-learn==0.23.1
optuna==1.4.0